+++ 13PeoplesSky/project2D/PursueBehaviour.cpp
@@ -19,7 +19,7 @@ Vector2 PursueBehaviour::update(Actor* pTarget, Enemy* pPursuer)
 {
 	Vector2 ToPlayer = pTarget->GetPosition() - pPursuer->GetPosition(); //Getting the distance between the pursuer and the target
 
-	double RelativeHeading = pPursuer->GetVelocity().dot(pTarget->GetVelocity());
+	float RelativeHeading = pPursuer->GetVelocity().dot(pTarget->GetVelocity());
 
 	//If the pursued target is facing us and travelling in our direction, then we don't need to pursue them, we can simply seek to their current position
 	// We calculate whether or not they are facing us if they're heading is within 18 degrees of our own
@@ -29,7 +29,7 @@ Vector2 PursueBehaviour::update(Actor* pTarget, Enemy* pPursuer)
 	}
 
 	//Here we are seeing where the pursued object is going to be in the future, and we will be seeking that location
-	double LookAheadTime = ToPlayer.magnitude() / (pPursuer->GetMaxSpeed() + pTarget->GetVelocity().magnitude());
+	float LookAheadTime = ToPlayer.magnitude() / (pPursuer->GetMaxSpeed() + pTarget->GetVelocity().magnitude());
 
 	return m_seek->update(pTarget->GetPosition() + pTarget->GetVelocity() * LookAheadTime, pPursuer);
 }
